requirements.txt
numpy
pillow
lxml
requests
flow_githubmanager @ git+https://github.com/The-Flow-Project/package-github-api@main
pydantic~=2.11.4
setuptools
PyYAML
htrflow
motor~=3.5.1